Keep event listener bound to element
Removing or adding an event listener causes a layout invalidation on
safari, which can lead to high latencyin scrolling on very large/complex
pages. Instead of addEventListener and removeEventListener on every
scroll start/stop, use a generic event listener function and a map that
keeps track of the actual handlers.